protect posts/ create, update, delete
Add authentication to posts and add user as author to new posts.
Check that the user owns the post before allowing update or delete.

+++ controllers/api/posts.js
@@ -1,8 +1,8 @@
-const Post = require('../../models/post')
+const Post = require('../../models/post');
 
 async function index(req, res) {
   try {
-    const posts = await Post.find({})
+    const posts = await Post.find({});
     res.json(posts);
   } catch (err) {
     res.status(400).json("database query failed");
@@ -11,7 +11,7 @@ async function index(req, res) {
 
 async function show(req, res) {
   try {
-    const post = await Post.findById(req.params.postId)
+    const post = await Post.findById(req.params.postId);
     res.status(200).json(post);
   } catch (err) {
     res.status(404).json("database query failed");
@@ -20,28 +20,47 @@ async function show(req, res) {
 
 async function create(req, res) {
   try {
-    const newPost = await Post.create(req.body)
+    req.body.author = req.user._id;
+    const newPost = await Post.create(req.body);
     res.status(200).json(newPost);
   } catch (err) {
-    res.status(400).json("database query failed");
+    res.status(400).json("post creation failed");
   }
 }
 
 async function update(req, res) {
   try {
-    const result = await Post.update({_id: req.params.postId}, req.body)
-    res.status(200).json(result);
+    let post = await Post.findById(req.params.postId);
+
+    // Here we have to cast the author._id object to a String before comparing
+    // We could also just add author to the query above, same result.
+    if (req.user._id !== String(post.author._id)) {
+      return res.status(403).json({
+        message: "Update failed. You are not the author of this post."
+      });
+    }
+
+    const result = await Post.updateOne({_id: req.params.postId}, req.body);
+    return res.status(200).json(result);
   } catch (err) {
-    res.status(400).json("database query failed");
+    return res.status(500).json("Update failed. Internal error.");
   }
 }
 
 async function _delete(req, res) {
   try {
-    const result = await Post.deleteOne({_id: req.params.postId})
-    res.status(200).json(result);
+    let post = await Post.findById(req.params.postId);
+
+    if (req.user._id !== String(post.author._id)) {
+      return res.status(403).json({
+        message: "Delete failed. You are not the author of this post."
+      });
+    }
+
+    const result = await Post.deleteOne({_id: req.params.postId});
+    return res.status(200).json(result);
   } catch (err) {
-    res.status(400).json("database query failed");
+    return res.status(500).json("Delete Failed. Internal Error.");
   }
 }
 
@@ -51,4 +70,4 @@ module.exports = {
   create,
   update,
   delete: _delete,
-}
+};
